MonsterMesh: fix battle balance (CPU parity) + damage-path hardening
Balance — a trained L50 was beating an untrained L60 in 1-3 hits:
- CPU opponents (daycare-peer `fight` and `fight 0`) were built with flat
  DV8 and ZERO stat-experience, while the player loads real DVs + accumulated
  stat-exp from the .sav (~+60% stats). Give CPU parity: max DVs + full
  stat-exp so foes aren't systematically frail. Tunable.
- isqrt16() seeded b at 2^15 (an odd power of two), misaligning the bit-search
  and inflating every stat-exp bonus by ~sqrt(2). Seed at 2^30 (power of four).
- mdata() took uint8_t but BattlePoke::moves are uint16 and Gen-3 move ids go
  to 354 — ids >255 truncated to the wrong move. Widen to uint16_t so Gen-3
  move data resolves correctly (everything runs the Gen-3 engine by default).

Damage-path security (audit findings):
- calcDamage divided by the defender's Def with no zero guard; a hostile wire
  party (or corrupt save) with def=0 caused a divide-by-zero reboot. Clamp
  D>=1 at both divide sites (main + confusion self-hit).
- WireParty validator checked species/level/moves but not the 6 stat fields;
  reject stat==0 (div-by-zero) or >999 (unreachable in Gen 1-3) at the trust
  boundary in both tbUnpackAndValidateWireParty and tbValidateWireParty. Also
  closes the all-0xFFFF "invincible one-shot mon" cheat on rated ladders.

Cleanup: drop unused DAILY_XP_CAP/XP_PER_LEVEL constants and correct the stale
"NOT wired into build yet" daycare comment.

Verified: battle 15/15, growth 4/4 (clang++ harness). Flashed to Red+Blue;
user confirms combat balance is correct.

+++ src/modules/monstermesh/Gen1BattleEngine.h
@@ -199,7 +199,7 @@ private:
     bool     isGen3() const { return gen_ >= 3; }
     // Move data for the active mechanics generation (gen3 re-types the
     // same 165 ids; effects+PP identical, so safe everywhere).
-    const Gen1MoveData *mdata(uint8_t id) const;
+    const Gen1MoveData *mdata(uint16_t id) const;
     uint16_t calcDamage(uint8_t side, uint8_t targetSide,
                         const Gen1MoveData &mv, bool &outCrit);
     uint8_t  effectiveness(uint8_t atkType, uint8_t defType) const;
